Moved Vertex struct to separate file vulkan_vertices header and cpp files in VulkanEngine library. Moved createVertexBuffer as well. Created  mapBufferMemory function for VulkanEngine class to copy data to buffers. Added BufferBundle struct to aggregate VkBuffer and VkDeviceMemory pair and added function to destroy these resources. Added struct vertexDecriptions to pass the vertices binding and attributes descriptions to the createGraphicsPipeline function. Made createBuffer and copyBuffer methods public for VulkanEngine. Moved writeCommandBuffer to main example code. Added vertexBuffer BufferBundle to Main application.
