+++ example/VulkanTutorial.cpp
@@ -1,7 +1,8 @@
 #include "vulkan_engine.h"
+#include "vulkan_vertices.h"
+#include "vulkan_imgui.h"
 #include "glfwInteraction.h"
 #include "importResources.h"
-#include "vulkan_imgui.h"
 #include "imgui.h"
 #include "imgui_impl_glfw.h"
 #include "imgui_impl_vulkan.h"
@@ -162,6 +163,7 @@ public:
 private:
 	shaderCode vert;
 	shaderCode frag;
+	BufferBundle vertexBuffer;
 
 	std::array<SampledImage, MAX_SAMPLED_IMAGES> textureImages;
 	std::array<SampledImage, MAX_SAMPLED_IMAGES> updatedTextureImages;
@@ -211,12 +213,12 @@ private:
 		createDescriptorSetLayout();
 		vert = readFile("./vert.spv");
 		frag = readFile("./frag.spv");
-		createGraphicsPipeline(vert, frag);
+		createGraphicsPipeline(vert, frag, Vertex::getDescriptions());
 		createCommandPool();
 		createColorResources();
 		createDepthResources();
 		createFramebuffers();
-		createVertexBuffer(vertices);
+		vertexBuffer = createVertexBuffer((VulkanEngine*)this, vertices);
 		createIndexBuffer(indices);
 		createCommandBuffers();
 		createUniformBuffers();
@@ -282,8 +284,7 @@ private:
 		vkDestroyBuffer(device, indexBuffer, nullptr);
 		vkFreeMemory(device, indexBufferMemory, nullptr);
 		// Here there is a choice for the Allocator function
-		vkDestroyBuffer(device, vertexBuffer, nullptr);
-		vkFreeMemory(device, vertexBufferMemory, nullptr);
+		destroyBufferBundle(vertexBuffer);
 
 		for (size_t i = 0; i < MAX_FRAME_IN_FLIGHT; i++) {
 			// Here there is a choice for the Allocator function
@@ -371,7 +372,7 @@ private:
 		// The render pass depends on the format of the swap chain. It is rare that the format changes but to be sure
 		createRenderPass();
 		createDescriptorPool();
-		createGraphicsPipeline(vert,frag);
+		createGraphicsPipeline(vert,frag, Vertex::getDescriptions());
 		createColorResources();
 		createDepthResources();
 		createFramebuffers();
@@ -693,6 +694,88 @@ private:
 		createSampledImage(image, matImage.cols, matImage.rows, matImage.elemSize(),(char*) matImage.data);
 
 		matImage.release();
+	}
+
+	void writeCommandBuffers() {
+
+		vkResetCommandPool(device, commandPool, VK_COMMAND_POOL_RESET_RELEASE_RESOURCES_BIT);
+		for (size_t i = 0; i < commandBuffers.size(); i++) {
+
+			VkCommandBufferBeginInfo beginInfo = {};
+			beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
+			//• VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT : The command buffer will be rerecorded right after 
+			//	executing it once.
+			//• VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT : This is a secondary command buffer that will be 
+			//	entirely within a single render pass.
+			//• VK_COMMAND_BUFFER_USAGE_SIMULTANEOUS_USE_BIT : The command buffer can be resubmitted while it is also 
+			//	already pending execution.
+			beginInfo.flags = 0; // OPtional
+			//  The pInheritanceInfo parameter is only relevant for secondary command
+			//	buffers.It specifies which state to inherit from the calling primary command
+			//	buffers.
+			beginInfo.pInheritanceInfo = nullptr; // Optional
+
+			if (vkBeginCommandBuffer(commandBuffers[i], &beginInfo) != VK_SUCCESS) {
+				throw std::runtime_error("failed to begin recording command buffer!");
+			}
+
+			VkRenderPassBeginInfo renderPassInfo = {};
+			renderPassInfo.sType = VK_STRUCTURE_TYPE_RENDER_PASS_BEGIN_INFO;
+			renderPassInfo.renderPass = renderPass;
+			renderPassInfo.framebuffer = swapChainFramebuffers[i];
+			renderPassInfo.renderArea.offset = { 0,0 };
+			renderPassInfo.renderArea.extent = swapChainExtent;
+
+			std::array<VkClearValue, 2> clearValues = {};
+			clearValues[0].color = { 0.0f,0.0f,0.0f,1.0f };
+			clearValues[1].depthStencil = { 1.0f, 0 };
+
+			//Note that the order of clearValues should be identical to the order of your attachments in the subpass. 
+			renderPassInfo.clearValueCount = static_cast<uint32_t>(clearValues.size());
+			renderPassInfo.pClearValues = clearValues.data();
+
+			//All of the functions that record commands can be recognized by their vkCmd prefix.
+
+			//• VK_SUBPASS_CONTENTS_INLINE : The render pass commands will be embedded in the primary command 
+			//	buffer itself and no secondary command buffers will be executed.
+			//• VK_SUBPASS_CONTENTS_SECONDARY_COMMAND_BUFFERS : The render pass commands will be executed
+			//	from secondary command buffers.
+			vkCmdBeginRenderPass(commandBuffers[i], &renderPassInfo, VK_SUBPASS_CONTENTS_INLINE);
+
+			vkCmdBindPipeline(commandBuffers[i], VK_PIPELINE_BIND_POINT_GRAPHICS, graphicsPipeline);
+
+			//• vertexCount : Even though we don’t have a vertex buffer, we technically
+			//	still have 3 vertices to draw.
+			//• instanceCount : Used for instanced rendering, use 1 if you’re not doing
+			//	that.
+			//• firstVertex : Used as an offset into the vertex buffer, defines the lowest
+			//	value of gl_VertexIndex.
+			//• firstInstance : Used as an offset for instanced rendering, defines the
+			//	lowest value of gl_InstanceIndex.
+			//vkCmdDraw(commandBuffers[i], 3, 1, 0, 0);
+			// EDIT: The previous commented was for a hard coded vertices in the shader. The application now allocates a vertex buffer outside
+			// and copys the vertices data from the application to the buffer
+
+			VkBuffer vertexBuffers[] = { vertexBuffer.buffer };
+			VkDeviceSize offsets[] = { 0 };
+			vkCmdBindVertexBuffers(commandBuffers[i], 0, 1, vertexBuffers, offsets);
+			// We can only have a single index buffer.
+			vkCmdBindIndexBuffer(commandBuffers[i], indexBuffer, 0, VK_INDEX_TYPE_UINT32);
+
+			// Descriptor Set bindings.ffs
+			vkCmdBindDescriptorSets(commandBuffers[i], VK_PIPELINE_BIND_POINT_GRAPHICS, pipelineLayout, 0, 1, &(descriptorSets[descriptorGroup][i]), 0, nullptr);
+			// Draw Vertex
+			//vkCmdDraw(commandBuffers[i], static_cast<uint32_t>(vertices.size()), 1, 0, 0);
+			// Indexed Vertex Draw
+			vkCmdDrawIndexed(commandBuffers[i], indexCount, 1, 0, 0, 0);
+
+
+			vkCmdEndRenderPass(commandBuffers[i]);
+
+			if (vkEndCommandBuffer(commandBuffers[i]) != VK_SUCCESS) {
+				throw std::runtime_error("failed to record command buffer!");
+			}
+		}
 	}
 };
