+++ include/vulkan_engine.h
@@ -18,8 +18,6 @@
 #include <glm\glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <glm/gtc/type_ptr.hpp>
-#define GLM_ENABLE_EXPERIMENTAL
-#include <glm/gtx/hash.hpp> // For Vertex hashing
 
 
 #include <vector>
@@ -63,57 +61,6 @@ typedef struct _SampledImage {
 	VkSampler sampler = VK_NULL_HANDLE;
 } SampledImage;
 
-struct Vertex {
-	alignas(16) glm::vec3 pos;
-	alignas(16) glm::vec3 color;
-	alignas(8) glm::vec2 texCoord;
-	alignas(4) int texId;
-
-
-	static VkVertexInputBindingDescription getBindingDescription() {
-		// Binding Description tells the rate at which vertex data is coming in
-		VkVertexInputBindingDescription bindingDescription = {};
-		bindingDescription.binding = 0;								// Binding index of this binding in the binding array
-		bindingDescription.stride = sizeof(Vertex);					// Stride between each vertex data 
-		// • VK_VERTEX_INPUT_RATE_VERTEX : Move to the next data entry after each vertex
-		// • VK_VERTEX_INPUT_RATE_INSTANCE : Move to the next data entry after each instance
-		bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-		return bindingDescription;
-	}
-	static std::array<VkVertexInputAttributeDescription, 4>getAttributeDescriptions() {
-		// This struct tells how to extract a vertex attribute from a vertex. Since we have two attributes, position and color, we have an array of two structs.
-		std::array<VkVertexInputAttributeDescription, 4>attributeDescriptions = {};
-		attributeDescriptions[0].binding = 0;
-		attributeDescriptions[0].location = 0;
-		//• float: VK_FORMAT_R32_SFLOAT
-		//• vec2 : VK_FORMAT_R32G32_SFLOAT
-		//• vec3 : VK_FORMAT_R32G32B32_SFLOAT
-		//• vec4 : VK_FORMAT_R32G32B32A32_SFLOAT
-		attributeDescriptions[0].format = VK_FORMAT_R32G32B32_SFLOAT;
-		attributeDescriptions[0].offset = offsetof(Vertex, pos);
-
-		attributeDescriptions[1].binding = 0;
-		attributeDescriptions[1].location = 1;
-		attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
-		attributeDescriptions[1].offset = offsetof(Vertex, color);
-
-		attributeDescriptions[2].binding = 0;
-		attributeDescriptions[2].location = 2;
-		attributeDescriptions[2].format = VK_FORMAT_R32G32_SFLOAT;
-		attributeDescriptions[2].offset = offsetof(Vertex, texCoord);
-
-
-		attributeDescriptions[3].binding = 0;
-		attributeDescriptions[3].location = 3;
-		attributeDescriptions[3].format = VK_FORMAT_R32_SINT;
-		attributeDescriptions[3].offset = offsetof(Vertex, texId);
-		return attributeDescriptions;
-	}
-
-	bool operator==(const Vertex& other) const {
-		return pos == other.pos && color == other.color && texCoord == other.texCoord;
-	}
-};
 
 struct UniformBufferObject {
 	alignas(16) glm::mat4 model;
@@ -173,6 +120,16 @@ struct SwapChainDetails {
 	std::vector<VkImageView> imageViews;
 };
 
+struct BufferBundle {
+	VkBuffer buffer;
+	VkDeviceMemory memory;
+};
+
+struct vertexDescriptions {
+	VkVertexInputBindingDescription binding;
+	std::vector<VkVertexInputAttributeDescription> attributes;
+};
+
 // *************** Other Functions *************
 void DestroyDebugUtilsMessengerEXT(VkInstance instance, VkDebugUtilsMessengerEXT debugMessenger, const VkAllocationCallbacks* pAllocator);
 
@@ -227,8 +184,6 @@ protected:
 	VkDeviceMemory colorImageMemory;
 	VkImageView colorImageView;
 
-	VkBuffer vertexBuffer;
-	VkDeviceMemory vertexBufferMemory;
 	VkBuffer indexBuffer;
 	VkDeviceMemory indexBufferMemory;
 	uint32_t indexCount = 0;
@@ -261,7 +216,7 @@ protected:
 	void createDescriptorSetLayout();
 	// For creating the graphics pipeline
 	VkShaderModule createShaderModule(const std::vector<char>& code);
-	void createGraphicsPipeline(shaderCode vert, shaderCode frag);
+	void createGraphicsPipeline(shaderCode vert, shaderCode frag, vertexDescriptions vertex);
 
 	// Various Resources
 	void createCommandPool();
@@ -271,7 +226,6 @@ protected:
 	void createFramebuffers();
 	void createSyncObjects();
 
-	void createVertexBuffer(std::vector<Vertex> vertices);
 	void createIndexBuffer(std::vector<uint32_t> indices);
 
 	// Descriptors
@@ -280,14 +234,9 @@ protected:
 	void createUniformBuffers();
 	void createSampledImage(SampledImage& image, int cols, int rows, int elemSize, char* imageData);
 	void updateDescriptorSet(std::array<SampledImage, MAX_SAMPLED_IMAGES> images, int groupIndex);
-	void writeCommandBuffers();
-
 	
 
 	uint32_t findMemoryType(uint32_t typeFilter, VkMemoryPropertyFlags properties);
-	void createBuffer(VkDeviceSize size, VkBufferUsageFlags usage,
-		VkMemoryPropertyFlags properties, VkBuffer& buffer, VkDeviceMemory& bufferMemory);
-	void copyBuffer(VkBuffer srcBuffer, VkBuffer dstBuffer, VkDeviceSize size);
 	void copyBufferToImage(VkBuffer buffer, VkImage image, uint32_t width, uint32_t height);
 	
 	void createImage(uint32_t width, uint32_t height,
@@ -308,6 +257,11 @@ protected:
 	void cleanupSampledImage(SampledImage& image);
 
 public:
+	void mapBufferMemory(VkDeviceMemory bufferMemory, void* data, VkDeviceSize datalen);
+	void destroyBufferBundle(BufferBundle buffer);
+	void createBuffer(VkDeviceSize size, VkBufferUsageFlags usage,
+		VkMemoryPropertyFlags properties, VkBuffer& buffer, VkDeviceMemory& bufferMemory);
+	void copyBuffer(VkBuffer srcBuffer, VkBuffer dstBuffer, VkDeviceSize size);
 	VkCommandBuffer beginSingleTimeCommands();
 	void endSingleTimeCommands(VkCommandBuffer commandBuffer);
